pyproject.toml
[project]
name = "polymarket-strategy-tester"
version = "1.0.0"
description = "Paper trading bot comparing undervalued vs momentum strategies on Polymarket 15-min BTC markets"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-asyncio>=0.21.0"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
